fix: infinite scroll formatting
