+++ frontend/src/components/profile/ProfileHeaderAndTabs.tsx
@@ -1,18 +1,38 @@
-import { Stack, Tabs, Tab, SxProps, Theme, useTheme } from '@mui/material';
+import {
+  Stack,
+  Tabs,
+  Tab,
+  SxProps,
+  Theme,
+  useTheme,
+  AppBar,
+  IconButton,
+  Toolbar,
+} from '@mui/material';
 import { User } from 'src/@types/user';
 import { capitalCase } from 'change-case';
 import ProfileHeader from './ProfileHeader';
 import useTabs from '../../hooks/ui/useTabs';
+import Iconify from '../Iconify';
+import { useNavigate } from 'react-router';
 
 interface ProfileProps {
   sx?: SxProps<Theme>;
+  showBack?: boolean;
   user: User;
   aboutTab: React.ReactElement;
   betasTab: React.ReactElement;
 }
 
-export default function ProfileHeaderAndTabs({ sx, user, aboutTab, betasTab }: ProfileProps) {
+export default function ProfileHeaderAndTabs({
+  sx,
+  user,
+  aboutTab,
+  betasTab,
+  showBack,
+}: ProfileProps) {
   const theme = useTheme();
+  const navigate = useNavigate();
   const { currentTab, onChangeTab } = useTabs('about');
 
   const PROFILE_TABS = [
@@ -31,22 +51,54 @@ export default function ProfileHeaderAndTabs({ sx, user, aboutTab, betasTab }: P
       <Stack
         spacing={2}
         direction="column"
-        sx={{ position: 'fixed', zIndex: theme.zIndex.appBar, pt: 4, px: 2, background: 'white', width: '100%' }}
+        sx={{
+          position: 'fixed',
+          zIndex: theme.zIndex.appBar,
+          left: 0,
+          top: 0,
+          background: 'white',
+          width: '100%',
+        }}
       >
-        <ProfileHeader user={user} />
-        <Tabs
-          allowScrollButtonsMobile
-          variant="scrollable"
-          scrollButtons="auto"
-          value={currentTab}
-          onChange={onChangeTab}
-        >
-          {PROFILE_TABS.map((tab) => (
-            <Tab disableRipple key={tab.value} value={tab.value} label={capitalCase(tab.value)} />
-          ))}
-        </Tabs>
+        {showBack && (
+          <AppBar
+            sx={{
+              background: 'white',
+              color: 'text.primary',
+              width: '100%',
+            }}
+            position="static"
+          >
+            <Toolbar sx={{ pl: 1 }}>
+              <IconButton
+                sx={{ mr: 1 }}
+                color="primary"
+                onClick={() => {
+                  navigate(-1);
+                }}
+              >
+                <Iconify icon="eva:arrow-back-fill" color="primary" />
+              </IconButton>
+            </Toolbar>
+          </AppBar>
+        )}
+        <Stack sx={{ px: 2, pt: showBack ? 2 : 4 }} direction="column" spacing={1}>
+          <ProfileHeader user={user} />
+          <Tabs
+            allowScrollButtonsMobile
+            variant="scrollable"
+            scrollButtons="auto"
+            value={currentTab}
+            onChange={onChangeTab}
+          >
+            {PROFILE_TABS.map((tab) => (
+              <Tab disableRipple key={tab.value} value={tab.value} label={capitalCase(tab.value)} />
+            ))}
+          </Tabs>
+        </Stack>
       </Stack>
-      <Stack sx={{ mt: '172px !important' }}>
+      <Stack sx={{ height: showBack ? '190px' : '140px' }} />
+      <Stack>
         {PROFILE_TABS.map((tab) => {
           const isMatched = tab.value === currentTab;
           return isMatched && tab.component;
